Show `InProgressBanner`.
