+++ assets/js/modules/analytics-4/components/dashboard/EnhancedMeasurementActivationBanner/index.js
@@ -26,35 +26,52 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import Data from 'googlesitekit-data';
+import { CORE_FORMS } from '../../../../../googlesitekit/datastore/forms/constants';
 import { CORE_MODULES } from '../../../../../googlesitekit/modules/datastore/constants';
 import { CORE_USER } from '../../../../../googlesitekit/datastore/user/constants';
-import { MODULES_ANALYTICS_4 } from '../../../datastore/constants';
 import {
+	ENHANCED_MEASUREMENT_ENABLED,
+	ENHANCED_MEASUREMENT_FORM,
+	MODULES_ANALYTICS_4,
+} from '../../../datastore/constants';
+import {
+	ACTIVATION_STEP_IN_PROGRESS,
 	ACTIVATION_STEP_SETUP,
 	ACTIVATION_STEP_SUCCESS,
 	ENHANCED_MEASUREMENT_ACTIVATION_BANNER_TOOLTIP_STATE_KEY,
 	ENHANCED_MEASUREMENT_ACTIVATION_BANNER_DISMISSED_ITEM_KEY,
 } from '../../../constants';
+import {
+	EDIT_SCOPE,
+	FORM_SETUP,
+} from '../../../../analytics/datastore/constants';
+import { ERROR_CODE_MISSING_REQUIRED_SCOPE } from '../../../../../util/errors';
 import { useTooltipState } from '../../../../../components/AdminMenuTooltip/useTooltipState';
 import { useShowTooltip } from '../../../../../components/AdminMenuTooltip/useShowTooltip';
 import { AdminMenuTooltip } from '../../../../../components/AdminMenuTooltip/AdminMenuTooltip';
+import InProgressBanner from './InProgressBanner';
 import SetupBanner from './SetupBanner';
 import SuccessBanner from './SuccessBanner';
-import { getTimeInSeconds } from '../../../../../util';
+import { getTimeInSeconds, trackEvent } from '../../../../../util';
 import whenActive from '../../../../../util/when-active';
 import {
 	isValidPropertyID,
 	isValidWebDataStreamID,
 } from '../../../utils/validation';
+import useViewContext from '../../../../../hooks/useViewContext';
 
 const { useSelect, useDispatch } = Data;
 
 function EnhancedMeasurementActivationBanner() {
+	const viewContext = useViewContext();
+
 	const [ step, setStep ] = useState( ACTIVATION_STEP_SETUP );
 	const [
 		isEnhancedMeasurementInitiallyDisabled,
 		setIsEnhancedMeasurementInitiallyDisabled,
 	] = useState( undefined );
+	const [ isEnabling, setIsEnabling ] = useState( false );
+	const [ errorNotice, setErrorNotice ] = useState( null );
 
 	const propertyID = useSelect( ( select ) =>
 		select( MODULES_ANALYTICS_4 ).getPropertyID()
@@ -90,7 +107,17 @@ function EnhancedMeasurementActivationBanner() {
 		);
 	} );
 
-	const { dismissItem } = useDispatch( CORE_USER );
+	const hasEditScope = useSelect( ( select ) =>
+		select( CORE_USER ).hasScope( EDIT_SCOPE )
+	);
+
+	const autoSubmit = useSelect( ( select ) =>
+		select( CORE_FORMS ).getValue( FORM_SETUP, 'autoSubmit' )
+	);
+
+	const { setValues } = useDispatch( CORE_FORMS );
+	const { dismissItem, setPermissionScopeError } = useDispatch( CORE_USER );
+	const { submitChanges } = useDispatch( MODULES_ANALYTICS_4 );
 
 	const { isTooltipVisible } = useTooltipState(
 		ENHANCED_MEASUREMENT_ACTIVATION_BANNER_TOOLTIP_STATE_KEY
@@ -110,11 +137,68 @@ function EnhancedMeasurementActivationBanner() {
 		);
 	}
 
-	const handleSubmit = useCallback( () => {
+	const commonSubmitChanges = useCallback( async () => {
+		setIsEnabling( true );
+
+		setValues( ENHANCED_MEASUREMENT_FORM, {
+			[ ENHANCED_MEASUREMENT_ENABLED ]: true,
+		} );
+
+		const { error } = await submitChanges();
+
+		setIsEnabling( false );
+
+		if ( error ) {
+			setErrorNotice( error );
+			return;
+		}
+
+		trackEvent(
+			`${ viewContext }_enhanced-measurement-notification`,
+			'confirm_notification'
+		);
+
 		if ( step === ACTIVATION_STEP_SETUP ) {
 			setStep( ACTIVATION_STEP_SUCCESS );
 		}
-	}, [ step ] );
+	}, [ setValues, submitChanges, viewContext, step ] );
+
+	const handleSubmitChanges = useCallback( async () => {
+		const scopes = [];
+
+		if ( hasEditScope === false ) {
+			scopes.push( EDIT_SCOPE );
+		}
+
+		// If scope not granted, trigger scope error right away. These are
+		// typically handled automatically based on API responses, but
+		// this particular case has some special handling to improve UX.
+		if ( scopes.length > 0 ) {
+			setValues( FORM_SETUP, { autoSubmit: true } );
+
+			setPermissionScopeError( {
+				code: ERROR_CODE_MISSING_REQUIRED_SCOPE,
+				message: __(
+					'Additional permissions are required to enable enhanced measurement for the selected web data stream.',
+					'google-site-kit'
+				),
+				data: {
+					status: 403,
+					scopes,
+					skipModal: true,
+					redirectURL: global.location.href,
+				},
+			} );
+			return;
+		}
+
+		await commonSubmitChanges();
+	}, [
+		hasEditScope,
+		commonSubmitChanges,
+		setPermissionScopeError,
+		setValues,
+	] );
 
 	useEffect( () => {
 		if (
@@ -127,6 +211,25 @@ function EnhancedMeasurementActivationBanner() {
 		isEnhancedMeasurementInitiallyDisabled,
 		isEnhancedMeasurementStreamEnabled,
 	] );
+
+	// If the user lands back on this component with autoSubmit and the edit scope,
+	// resubmit the form.
+	useEffect( () => {
+		async function handleAutoSubmit() {
+			// Auto-submit should only auto-invoke once.
+			setValues( FORM_SETUP, { autoSubmit: false } );
+
+			if ( step === ACTIVATION_STEP_SETUP ) {
+				setStep( ACTIVATION_STEP_IN_PROGRESS );
+			}
+
+			await commonSubmitChanges();
+		}
+
+		if ( autoSubmit && hasEditScope ) {
+			handleAutoSubmit();
+		}
+	}, [ hasEditScope, setValues, commonSubmitChanges, autoSubmit, step ] );
 
 	if ( isTooltipVisible ) {
 		return (
@@ -154,10 +257,16 @@ function EnhancedMeasurementActivationBanner() {
 	if ( step === ACTIVATION_STEP_SETUP ) {
 		return (
 			<SetupBanner
-				onSubmitSuccess={ handleSubmit }
+				errorNotice={ errorNotice }
+				isSaving={ isEnabling }
 				onDismiss={ handleDismiss }
+				onSubmit={ handleSubmitChanges }
 			/>
 		);
+	}
+
+	if ( step === ACTIVATION_STEP_IN_PROGRESS ) {
+		return <InProgressBanner />;
 	}
 
 	if ( step === ACTIVATION_STEP_SUCCESS ) {
